Refactor AJAX call for single user/gym GET request
Prior to these edits, there were two nearly-identical AJAX calls for the GET
request for an individual gym and user resource. this caused problems - because
of event delegation, the button click to load either resource triggered both
AJAX calls, and obviously only one would return the desired data. This checks
to see which environment triggered the event and uses the same AJAX request for
either.