Creates gym array and update bulletins for climbs
Small change - gyms is now an array, which paves the way for future "subscription"
model of gym following. Current app still hardcodes newsfeed to always looks at
gym #2, if it exists. Needless to say that's not ideal.

Bigger change - upon sign in, gyms are set for the user. When bulletins are
displayed, a loop starts that starts its own loop for each item, checking to see
if any of the climbs from a given gym match the bulletin being displayed. If so,
that climb is rendered within that bulletin. Editing has been reintroduced by
simply clicking on the climb square.

Lots more refactoring needed, but I'm happy with it as is, for now.