Finish first round of refactoring
Index.js is now no longer like 900 lines long, which was the original goal. In the pursuit of
that goal, a bunch of new files were created that handle a specific set of functions. A few
are in use as catch-all handlers - 'event-handler.js', which will likely remain so, and
'ajax.js', which needs to be refactored further such that its constituent functions
live in more appropriate role-based files. Further refactoring needs to be done:
specifically, cleaning up this code so the calls are more semantic and less 'tangled',
with each file requiring a network of other files, and making the whole thing mobile optimized.